General: move VaultPress class to own file & add PHP53 shortcircuit
Fixes #15

- Moving the main VaultPress class to its own class allows us to move all PHP 5.3+ code out of the plugin's entry point, vaultpress.php
- In that plugin's entry point, we now check the PHP version in use on the site, and pause the VaultPress plugin (i.e. stop loading it) if you are running an old version of PHP.
